Create Decimal to binary counter
This assignment is done with C, It takes a users input integer value and converts it into a binary number. Then it will take the binary array and count the number of 1s and 0s, returning the number of 1s and 0s. Also whether 1 is an even or odd number as a result